+++ src/main/java/de/sakurajin/sakuralib/loot/v1/LootDistributionHelper.java
@@ -1,7 +1,8 @@
 package de.sakurajin.sakuralib.loot.v1;
 
+import de.sakurajin.sakuralib.loot.v2.distribution.PowerDistribution;
+
 import java.util.ArrayList;
-import java.util.HashMap;
 
 /**
  * This class can be used to create weight distributions for loot tables.
@@ -9,26 +10,16 @@ import java.util.HashMap;
  * * power distribution (@see LootDistributionHelper#getPowerDistribution(int, int))
  * <p>
  * Other distributions may be added in the future.
+ * @deprecated use the individual distributions directly instead. Will be removed in 2.0.0
+ * @see PowerDistribution for the v2 implementation
  */
+@Deprecated(since = "1.4.0", forRemoval = true)
 public class LootDistributionHelper {
-    /**
-     * This method contains a cache for previously calculated power distributions.
-     * The cache is structured as follows:
-     * * the first key is the order of the distribution
-     * * the second key is the number of elements in the distribution
-     * * the third is an array with the values of the distribution
-     */
-    static final HashMap<Integer, HashMap<Integer, ArrayList<Double>>> powerDistributionCache = new HashMap<>();
 
     /**
      * This method returns a power distribution for the given order and element count.
-     * The distribution is normalized and the sum of all elements is roughly 1.
-     * The distribution is cached, so it is not necessary to cache it yourself.
-     * <p>
-     * It is calculated using the following formula:
-     * * totalWeight = sum(i = 0, elementCount - 1, (elementCount - i) ^ order)
-     * * elementWeight(i) = (elementCount - i) ^ order / totalWeight
-     * The totalWeight is used to normalize the distribution.
+     * It passes the value to the v2 implementation.
+     * @see PowerDistribution for more information.
      *
      * @param order        the order of the distribution
      * @param elementCount the number of elements in the distribution
@@ -36,27 +27,7 @@ public class LootDistributionHelper {
      * @throws IllegalArgumentException if elementCount is less than or equal to 0
      */
     public static ArrayList<Double> getPowerDistribution(int order, int elementCount) throws IllegalArgumentException {
-        if (elementCount <= 0) {
-            throw new IllegalArgumentException("elementCount must be greater than 0");
-        }
-
-        var orderMap   = powerDistributionCache.getOrDefault(order, new HashMap<>());
-        var elementMap = orderMap.getOrDefault(elementCount, new ArrayList<>());
-
-        if (elementMap.isEmpty()) {
-            double totalWeight = 0;
-            for (int i = 0; i < elementCount; i++) {
-                totalWeight += Math.pow(elementCount - i, order);
-            }
-
-            for (int i = 0; i < elementCount; i++) {
-                elementMap.add(Math.pow(elementCount - i, order) / totalWeight);
-            }
-            orderMap.put(elementCount, elementMap);
-            powerDistributionCache.put(order, orderMap);
-        }
-
-        return elementMap;
+        return PowerDistribution.GetDistribution(order, elementCount);
     }
 
 }