+++ src/main/java/de/sakurajin/sakuralib/loot/v2/table_insert/LootEntryInsert.java
@@ -1,14 +1,29 @@
-package de.sakurajin.sakuralib.loot.v1;
+package de.sakurajin.sakuralib.loot.v2.table_insert;
 
+import de.sakurajin.sakuralib.loot.v2.LootSourceHelper;
 import net.minecraft.loot.entry.LootPoolEntry;
 
 /**
  * A class that represents an entry that should be inserted into a loot table.
  * It contains the entry that should be inserted, the index of the pool it should be inserted into and the loot sources
- * @deprecated Use {@link de.sakurajin.sakuralib.loot.v2.table_insert.LootEntryInsert} instead.
  */
-@Deprecated(since = "1.4.0", forRemoval = true)
-public class LootEntryInsert extends de.sakurajin.sakuralib.loot.v2.table_insert.LootEntryInsert {
+public class LootEntryInsert {
+    /**
+     * The entry that should be inserted into a loot table.
+     */
+    public final LootPoolEntry entry;
+
+    /**
+     * The index of the pool the entry should be inserted into.
+     */
+    public final int poolIndex;
+
+    /**
+     * The loot sources that should be used for the insertion.
+     *
+     * @see LootSourceHelper for more information on how this integer works.
+     */
+    public final int lootSources;
 
     /**
      * Creates a new LootEntryInsert.
@@ -18,7 +33,9 @@ public class LootEntryInsert extends de.sakurajin.sakuralib.loot.v2.table_insert
      * @param lootSources The loot sources that should be used for the insertion.
      */
     public LootEntryInsert(LootPoolEntry entry, int poolIndex, int lootSources) {
-        super(entry, poolIndex, lootSources);
+        this.entry       = entry;
+        this.poolIndex   = poolIndex;
+        this.lootSources = lootSources;
     }
 
     /**
@@ -28,7 +45,7 @@ public class LootEntryInsert extends de.sakurajin.sakuralib.loot.v2.table_insert
      * @param poolIndex The index of the pool the entry should be inserted into.
      */
     public LootEntryInsert(LootPoolEntry entry, int poolIndex) {
-        super(entry, poolIndex, LootSourceHelper.BUILTIN);
+        this(entry, poolIndex, LootSourceHelper.BUILTIN);
     }
 
     /**
@@ -37,6 +54,6 @@ public class LootEntryInsert extends de.sakurajin.sakuralib.loot.v2.table_insert
      * @param entry The entry that should be inserted into a loot table.
      */
     public LootEntryInsert(LootPoolEntry entry) {
-        super(entry, 0, LootSourceHelper.BUILTIN);
+        this(entry, 0, LootSourceHelper.BUILTIN);
     }
 }